webapp/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gitpython>=3.1.43
pydantic>=2.9.0
python-dotenv>=1.0.1
httpx>=0.27.0
aiofiles>=23.2.1
sqlalchemy>=2.0.0
openai>=1.30.0
